+++ src/utils/rgb.c
@@ -6,13 +6,13 @@
 /*   By: hlesny <hlesny@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 00:08:53 by srapin            #+#    #+#             */
-/*   Updated: 2023/11/29 03:23:46 by hlesny           ###   ########.fr       */
+/*   Updated: 2023/11/29 17:54:26 by hlesny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/mini_rt.h"
 
-unsigned long	create_rgb(unsigned char r, unsigned char g, unsigned char b)
+unsigned int	create_rgb(unsigned char r, unsigned char g, unsigned char b)
 {
 	t_color	c;
 
@@ -23,7 +23,7 @@ unsigned long	create_rgb(unsigned char r, unsigned char g, unsigned char b)
 	return (c.hex);
 }
 
-bool	get_rgb(char *str, unsigned long *c)
+bool	get_rgb(char *str, unsigned int *c)
 {
 	int i = 0;
 	bool flag;
